Fixed a lot of warnings added by -Wextra and more flags. There are more remaining.

+++ graphics/screen.cpp
@@ -175,9 +175,9 @@ bool Screen::setVideoMode()
 {
 	Settings *config = Settings::pInstance();
 
-	unsigned int ScreenWidth  = config->get("ScreenWidth")->toInt();
-	unsigned int ScreenHeight = config->get("ScreenHeight")->toInt();
-	unsigned int ScreenBpp    = config->get("ScreenBpp")->toInt();
+	unsigned int ScreenWidth  = (unsigned int) config->get("ScreenWidth")->toInt();
+	unsigned int ScreenHeight = (unsigned int) config->get("ScreenHeight")->toInt();
+	unsigned int ScreenBpp    = (unsigned int) config->get("ScreenBpp")->toInt();
 	bool Fullscreen  = config->get("Fullscreen")->toBool();
 
 	if (m_screen != NULL)
@@ -332,7 +332,7 @@ void Screen::resetViewport()
 		int weight = int((m_ratio/screen_ratio)*(float)m_selectedMode.w);
 		int x = (m_selectedMode.w - weight)/2;
 		glViewport(x,0,weight,m_selectedMode.h);
-		//glScissor (x,0,weight,mySelectedMode.h);
+		//glScissor (x,0,weight,m_selectedMode.h);
 		//glEnable(GL_SCISSOR_TEST);
 	}
 	else if ( screen_ratio < m_ratio)
@@ -340,7 +340,7 @@ void Screen::resetViewport()
 		int height = int((screen_ratio/m_ratio)*(float)m_selectedMode.h);
 		int y = (m_selectedMode.h - height)/2;
 		glViewport(0,y,m_selectedMode.w,height);
-		//glScissor (0,y,mySelectedMode.w,height);
+		//glScissor (0,y,m_selectedMode.w,height);
 		//glEnable(GL_SCISSOR_TEST);
 	}
 	else
